Move BINDING nodes + BINDS edges into enhacements
