+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/standard/MetaDataTests.scala
@@ -1,9 +1,9 @@
-package io.shiftleft.fuzzyc2cpg.querying
+package io.shiftleft.fuzzyc2cpg.standard
 
 import io.shiftleft.fuzzyc2cpg.testfixtures.FuzzyCCodeToCpgSuite
 import io.shiftleft.semanticcpg.language._
 
-class MetaDataNodeTests extends FuzzyCCodeToCpgSuite {
+class MetaDataTests extends FuzzyCCodeToCpgSuite {
 
   override val code: String =
     """
