+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/standard/ParameterTests.scala
@@ -1,0 +1,3 @@
+package io.shiftleft.fuzzyc2cpg.standard
+
+class ParameterTests {}
