+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/standard/TypeDeclTests.scala
@@ -1,4 +1,4 @@
-package io.shiftleft.fuzzyc2cpg.querying
+package io.shiftleft.fuzzyc2cpg.standard
 
 import io.shiftleft.fuzzyc2cpg.testfixtures.FuzzyCCodeToCpgSuite
 import io.shiftleft.semanticcpg.language._
